tests/wpt/web-platform-tests/css/css-regions/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Regions properties</title>
<link rel="help" href="https://drafts.csswg.org/css-regions/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_not_inherited('flow-from', 'none', 'foo');
assert_not_inherited('flow-into', 'none', 'bar');
assert_not_inherited('region-fragment', 'auto', 'break');
</script>
</body>
</html>
